upgrade asyncHttpClient from ning (1.x) to 2.x branch of 2.6.0. refactor package name to drop ning
