+++ build.gradle
@@ -82,7 +82,7 @@ ext {
     objenesisVersion = '2.1'
     javassistVersion = '3.18.2-GA'
     jacksonVersion = '2.8.9'
-    ningAsyncHttpClientVersion = '1.9.38'
+    asyncHttpClientVersion = '2.6.0'
     guiceVersion = '4.0'
 
     jodaTimeVersion = '2.9.2'
